Use a clock-based eviction strategy
Instead of queueing potential pages to evict when unpinning, all pages currently in memory are kept in the evictionqueue.
When more space is needed, pages are evicted if they haven't changed since the last time we scanned them.

Based on the hash table implementation in the VMCache paper, except that
we store page states by reference and for simplicity store the eviction
candidates in a circular buffer instead of a hash table since lookups
aren't necessary.